FIX: use unsigned int for image dimensions
Using uint64_t was causing errors with wrapping, so for now we switch to using
unsigned int for size. The errors were only occuring with std::vector< uint64_t
>

+++ Testing/Unit/ImageTests.cxx
@@ -101,7 +101,7 @@ TEST_F(Image,Create) {
   EXPECT_EQ ( shortImage->GetSize()[1], shortImage->GetHeight() );
   EXPECT_EQ ( shortImage->GetSize()[2], shortImage->GetDepth() );
 
-  std::vector< uint64_t > size = shortImage->GetSize();
+  std::vector< unsigned int > size = shortImage->GetSize();
 
   EXPECT_EQ ( size[0], shortImage->GetWidth() );
   EXPECT_EQ ( size[1], shortImage->GetHeight() );
